the most basic of inheritance
pretty much just took everything from what was the particle object and moved it over to a parent confetti class and am going to use that for my different particles to inherit from, circles currently inherits everything from confetti as my first test using inheritance
